+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './globalNotifications/globalNotifications.component';
+export * from './notificationMessage/notificationMessage.component';
+export * from './notifications/notifications.component';