+++ src/directives/index.ts
@@ -1,0 +1,3 @@
+export * from './localNotificationsWithProvider/localNotificationsWithProvider.directive';
+export * from './messageOptionsNotifications/messageOptionsNotifications.directive';
+export * from './messageRenderer/messageRenderer.directive';