+++ src/modules/index.ts
@@ -1,0 +1,3 @@
+export * from './notificationsCore/notificationsCore.module';
+export * from './notificationsGlobal/notificationsGlobal.module';
+export * from './notificationsLocal/notificationsLocal.module';