+++ src/modules/notificationsGlobal/notificationsGlobal.module.ts
@@ -1,0 +1,21 @@
+import {NgModule} from '@angular/core';
+
+import {GlobalNotificationsComponent} from '../../components';
+
+/**
+ * Module for global notifications
+ */
+@NgModule(
+{
+    imports:
+    [
+        GlobalNotificationsComponent,
+    ],
+    exports:
+    [
+        GlobalNotificationsComponent,
+    ],
+})
+export class NotificationsGlobalModule
+{
+}