+++ src/modules/notificationsLocal/notificationsLocal.module.ts
@@ -1,0 +1,26 @@
+import {NgModule} from '@angular/core';
+
+import {NotificationsComponent} from '../../components';
+import {LocalNotificationsProviderDirective, MessageOptionsNotificationsDirective} from '../../directives';
+
+/**
+ * Module for local notifications
+ */
+@NgModule(
+{
+    imports:
+    [
+        NotificationsComponent,
+        LocalNotificationsProviderDirective,
+        MessageOptionsNotificationsDirective,
+    ],
+    exports:
+    [
+        NotificationsComponent,
+        LocalNotificationsProviderDirective,
+        MessageOptionsNotificationsDirective,
+    ],
+})
+export class NotificationsLocalModule
+{
+}